with diagram
