+++ Core/key_generator.py
@@ -1,0 +1,7 @@
+import os
+import binascii
+
+
+def generate_key():
+    key = binascii.hexlify(os.urandom(4))
+    return str(key)[2:-1]
