+++ Tests/test_ask_name.py
@@ -1,0 +1,6 @@
+from unittest import TestCase
+
+
+class TestAsk_name(TestCase):
+    def test_ask_name(self):
+        self.fail()
