--- config.py
@@ -1,8 +1,0 @@
-#token = '420182059:AAGF7cwnrfm8FObtZS4RPmU4gVvWuVQZ2Ug'
-# ILMK
-token = '342086039:AAE-rBthwuwaUlHH0Xbea6ZnU5Zw7zJ05X4'
-
-db_host = '37.46.132.57'
-db_username = 'telebot'
-db_password = 'Malinka2017'
-db_name = 'lmstelebot'
