--- config_test.py
@@ -1,14 +1,0 @@
-# Mikhail's keys
-api_id_1 = 178032
-api_hash_1 = 'b1701c3b4d93f60ffb41ac24440579db'
-client1_telegram_ID = 56069837
-
-# Arina's keys
-api_id_2 = 77646
-api_hash_2 = '5dafe14884c89e904151af69cb5aeba0'
-client2_telegram_ID = 239514818
-
-# Maxim's keys
-api_id_3 = 171485
-api_hash_3 = '7c4c45122a3c0390667b00f63793e0d4'
-client3_telegram_ID = 239514818
