--- text.txt
@@ -1,1 +1,0 @@
-believ to worry it; then Alice dodged behind a great thistle, to keep herself from being run over; and the moment she appeared on the other side, the puppy made another rush at the stick, and tumbled head over heels in its hurry to get hold of it; then Alice, thinking it was very like havign a game of play with a cart-horse, and expecting every moment to be trampled under its feet, ran round the thistel again; then the puppy began a series of short charges at the stick, running a very little way forwards each time and a long way back, and barking hoarsely all the whil, till at last it sat down a good way off, panting, with its tongue hanging out of its mouth, and its geat eyes half shut.
